k save before termination condition fix, last ish
